+++ .vscode/settings.json
@@ -5,6 +5,7 @@
     "files.associations": {
         "*.h": "c",
         "*.c": "c",
+        "cstdlib": "c",
         "*.md": "markdown",
         "*.sh": "shellscript",
         "*.cpp": "cpp",
