+++ src/CMakeLists.txt
@@ -5,6 +5,16 @@ cmake_minimum_required(VERSION 3.13)
 set(CMAKE_C_STANDARD 11)
 set(CMAKE_CXX_STANDARD 17)
 
+if (DEFINED ENV{AURORA_SRC_PATH} AND (NOT AURORA_SRC_PATH))
+    set(AURORA_SRC_PATH $ENV{AURORA_SRC_PATH})
+    message("Using AURORA_SRC_PATH from environment ('${AURORA_SRC_PATH}')")
+endif ()
+
+if (NOT DEFINED AURORA_SRC_PATH)
+    set(AURORA_SRC_PATH ${CMAKE_CURRENT_LIST_DIR})
+    message("Using default AURORA_SRC_PATH ('${AURORA_SRC_PATH}')")
+endif ()
+
 if (DEFINED ENV{PICO_BOARD} AND (NOT PICO_BOARD))
     set(PICO_BOARD $ENV{PICO_BOARD})
     message("Using PICO_BOARD from environment ('${PICO_BOARD}')")
@@ -18,11 +28,11 @@ include(pico_sdk_import.cmake)
 
 # Pull in FreeRTOS (must be before project)
 if (PICO_BOARD STREQUAL "pico2_w" OR PICO_BOARD STREQUAL "pico2")
-    set( BOARD_SRC_PATH ${CMAKE_CURRENT_LIST_DIR}/board/rp2350_arm CACHE STRING "Board path")
+    set( BOARD_SRC_PATH ${AURORA_SRC_PATH}/board/rp2350_arm CACHE STRING "Board path")
 endif()
 
 if (PICO_BOARD STREQUAL "pico_w" OR PICO_BOARD STREQUAL "pico")
-    set( BOARD_SRC_PATH ${CMAKE_CURRENT_LIST_DIR}/board/rp2040 CACHE STRING "Board path")
+    set( BOARD_SRC_PATH ${AURORA_SRC_PATH}/board/rp2040 CACHE STRING "Board path")
 endif ()
 
 if (NOT DEFINED BOARD_SRC_PATH)
@@ -36,6 +46,8 @@ if (PICO_SDK_VERSION_STRING VERSION_LESS "1.4.0")
 endif()
 
 project(aurora C CXX ASM)
+
+add_subdirectory(tasks)
 
 # Initialise the Raspberry Pi Pico SDK
 pico_sdk_init()
@@ -54,14 +66,17 @@ pico_enable_stdio_usb(aurora 1)
 target_link_libraries(aurora
         FreeRTOS-Kernel
         FreeRTOS-Kernel-Heap4
+        aurora_tasks
         pico_stdlib)
 
 # Add the standard include files to the build
-target_include_directories(aurora PRIVATE
-        ${CMAKE_CURRENT_LIST_DIR}
+set(AURORA_INCLUDES
         ${FREERTOS_KERNEL_PATH}/Common/include
         ${BOARD_SRC_PATH}/include
+        ${AURORA_SRC_PATH}/include
         )
+
+target_include_directories(aurora PRIVATE ${AURORA_INCLUDES})
 
 target_link_libraries(aurora
         hardware_i2c
@@ -69,4 +84,3 @@ target_link_libraries(aurora
         )
 
 pico_add_extra_outputs(aurora)
-
