+++ src/tasks/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_library(aurora_tasks STATIC
+        freertos_scheduling.c
+        watchdog_service.c
+        )
+
+target_link_libraries(aurora_tasks
+        FreeRTOS-Kernel
+        pico_stdlib)
+
+target_include_directories(aurora_tasks PRIVATE
+        ${AURORA_SRC_PATH}/include
+        ${FREERTOS_KERNEL_PATH}/Common/include
+        ${BOARD_SRC_PATH}/include
+)
