+++ src/tasks/freertos_scheduling.c
@@ -1,0 +1,72 @@
+/**
+ * @file freertos_scheduling.c
+ * @brief FreeRTOS basic scheduling tasks source file
+ * @note This file contains the function implementation for the freertos
+ * scheduler.
+ *
+ * Author: Maximilian Stephan @ Auxspace e.V.
+ * Copyright (C) 2025 Auxspace e.V.
+ */
+
+#include "FreeRTOS.h"
+#include "task.h"
+#include "queue.h"
+
+#include <stdio.h>
+
+#include <aurora/task/freertos_scheduling.h>
+
+/*----------------------------------------------------------------------------*/
+
+void vApplicationMallocFailedHook(void)
+{
+    /* Called if a call to pvPortMalloc() fails because there is insufficient
+    free memory available in the FreeRTOS heap.  pvPortMalloc() is called
+    internally by FreeRTOS API functions that create tasks, queues, software
+    timers, and semaphores.  The size of the FreeRTOS heap is set by the
+    configTOTAL_HEAP_SIZE configuration constant in FreeRTOSConfig.h. */
+
+    /* Force an assert. */
+    configASSERT((volatile void*)NULL);
+}
+
+/*----------------------------------------------------------------------------*/
+
+void vApplicationStackOverflowHook(TaskHandle_t pxTask, char* pcTaskName)
+{
+    (void)pcTaskName;
+    (void)pxTask;
+
+    /* Run time stack overflow checking is performed if
+    configCHECK_FOR_STACK_OVERFLOW is defined to 1 or 2.  This hook
+    function is called if a stack overflow is detected. */
+
+    /* Force an assert. */
+    configASSERT((volatile void*)NULL);
+}
+
+/*----------------------------------------------------------------------------*/
+
+void vApplicationIdleHook(void)
+{
+    volatile size_t xFreeHeapSpace;
+
+    /* This is just a trivial example of an idle hook.  It is called on each
+    cycle of the idle task.  It must *NOT* attempt to block.  In this case the
+    idle task just queries the amount of FreeRTOS heap that remains.  See the
+    memory management section on the http://www.FreeRTOS.org web site for memory
+    management options.  If there is a lot of heap memory free then the
+    configTOTAL_HEAP_SIZE value in FreeRTOSConfig.h can be reduced to free up
+    RAM. */
+    xFreeHeapSpace = xPortGetFreeHeapSize();
+
+    /* Remove compiler warning about xFreeHeapSpace being set but never used. */
+    (void)xFreeHeapSpace;
+}
+
+/*----------------------------------------------------------------------------*/
+
+void vApplicationTickHook(void)
+{
+    /* Do nothing ATM */
+}
